% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{collections}
\alias{collections}
\title{Load package collection}
\usage{
collections(x)
}
\arguments{
\item{x}{name of collection.}
}
\value{
Nothing, packages are silenty loaded.
}
\description{
Load a collection of packages.
}

