% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{corner}
\alias{corner}
\title{Return top or bottom left corner of object}
\usage{
corner(x, n = 5, head = TRUE)
}
\arguments{
\item{x}{same as for head}

\item{n}{same as for head}

\item{head}{logical: TRUE for head, FALSE for tail.}
}
\value{
same as for head/tail, but with the the same number of columns as rows.
}
\description{
Simple wrapper of head/tail that only prints the first few columns.
}
\examples{
corner(iris)
}
