% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_profile.R
\name{create_Rprofile}
\alias{create_Rprofile}
\title{Create .Rprofile}
\usage{
create_Rprofile()
}
\description{
Copy a template .Rprofile to the home directory
}
\details{
ThodbergMisc/extdata contains a template for an .Rprofile file for an effective R-setup that does not influence code execution. This file will try to adapt startup settings to whether R is run from a terminal or RStudio.
}

