% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{insertReverseAssignment}
\alias{insertReverseAssignment}
\title{Insert \%->\%.}
\usage{
insertReverseAssignment()
}
\description{
Call this function as an addin to insert \code{\%->\%} at the cursor position.
}

