% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{package_collections}
\alias{package_collections}
\title{Collections of R-packages}
\format{
A named list with package names
}
\usage{
package_collections
}
\description{
Various collections of useful R-packages.
}
\keyword{datasets}
