% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{run_loc}
\alias{run_loc}
\title{Detect whether the script is run locally or on server}
\usage{
run_loc(servername = "nzl922", mountpoint = "~/server")
}
\arguments{
\item{servername}{username on server.}

\item{mountpoint}{Path to sshfs mountpoint.}
}
\value{
Appropriate prefix for file locations on server.
}
\description{
This function detects via username whether the script is run from a server or locally with the server system mounted.
This function basically returns the mountpoint string if the user (as detected by Sys.info) is not servername.
This is useful when constructing file paths, since the same code will still get correct files if mounted via sshfs.
}
