% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{silent_library}
\alias{silent_library}
\title{Silently add packages}
\usage{
silent_library(packages)
}
\arguments{
\item{packages}{character vector of package names}
}
\description{
Load packages without any info returned
}
\details{
Uses invisible() and supressPackageStartMessages() to silently load the packages.
}

