% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{mount}
\alias{mount}
\title{Detect whether the script is run locally or on server}
\usage{
mount(server = c("hulk", "porus01"), mountpoint = "~/porus")
}
\arguments{
\item{server}{username on server.}

\item{mountpoint}{Path to sshfs mountpoint.}
}
\value{
Appropriate prefix for file locations on server.
}
\description{
This function detects via nodename whether the script is run from a server or locally with the server system mounted.
This function basically returns the mountpoint string if the user (as detected by Sys.info) is not a server.
This is useful when constructing file paths, since the same code will still get correct files if mounted via sshfs.
}
